test/fix out of date dir case
Sometimes dump() might return null, which necessitates
to re-acquire dir via an ls() call. Borrow checker
seems to attribute the lifetime even to None/Err option,
which results in a complile error.

Attempt to work around that by using directly the client_ptr
it is the only thing we need, anyway.